requirements.txt
python-telegram-bot==13.1
emoji==1.2.0
selenium
pyyaml
pyvirtualdisplay